requirements.txt
langchain-google-genai
langchain
python-dotenv
fastapi
uvicorn
python-multipart
